Separated login and registration cards and added them to the product page.
