+++ client/src/App.jsx
@@ -1,4 +1,4 @@
-import ProductPage from "./components/ProductPage"
+import ProductPage from "./Pages/ProductPage"
 
 function App() {
