+++ client/src/components/LoginCard.jsx
@@ -1,0 +1,134 @@
+
+import React from 'react';
+import { X, Mail, Lock, ArrowRight, AlertCircle } from 'lucide-react';
+import { Formik, Form, Field } from 'formik';
+import * as Yup from 'yup';
+
+const LoginSchema = Yup.object().shape({
+  email: Yup.string()
+    .email('Invalid email address')
+    .required('Email is required'),
+  password: Yup.string()
+    .min(8, 'Password must be at least 8 characters')
+    .required('Password is required'),
+  rememberMe: Yup.boolean()
+});
+
+const LoginCard = ({ onClose, switchToRegister }) => (
+  <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center">
+    <div className="bg-white rounded-lg p-8 w-[400px] relative shadow-2xl">
+      <button
+        onClick={onClose}
+        className="absolute top-4 right-4 text-gray-500 hover:text-gray-700 transition-colors"
+      >
+        <X size={24} />
+      </button>
+      
+      <div className="text-center mb-8">
+        <h2 className="text-3xl font-bold text-gray-900 mb-2">Welcome Back</h2>
+        <p className="text-gray-600">Please enter your details to sign in</p>
+      </div>
+      
+      <Formik
+        initialValues={{
+          email: '',
+          password: '',
+          rememberMe: false
+        }}
+        validationSchema={LoginSchema}
+        onSubmit={(values, { setSubmitting }) => {
+          setTimeout(() => {
+            alert(JSON.stringify(values, null, 2));
+            setSubmitting(false);
+          }, 400);
+        }}
+      >
+        {({ errors, touched, isSubmitting }) => (
+          <Form className="space-y-6">
+            <div className="relative">
+              <label className="block text-sm font-medium text-gray-700 mb-1">
+                Email
+              </label>
+              <div className="relative">
+                <Mail className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400" size={20} />
+                <Field
+                  name="email"
+                  type="email"
+                  className={`w-full pl-10 pr-4 py-3 border rounded-lg focus:ring-2 focus:ring-black focus:outline-none ${
+                    errors.email && touched.email ? 'border-red-500' : 'border-gray-300'
+                  }`}
+                  placeholder="Enter your email"
+                />
+              </div>
+              {errors.email && touched.email && (
+                <div className="flex items-center mt-1 text-red-500 text-sm">
+                  <AlertCircle size={16} className="mr-1" />
+                  {errors.email}
+                </div>
+              )}
+            </div>
+            
+            <div className="relative">
+              <label className="block text-sm font-medium text-gray-700 mb-1">
+                Password
+              </label>
+              <div className="relative">
+                <Lock className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400" size={20} />
+                <Field
+                  name="password"
+                  type="password"
+                  className={`w-full pl-10 pr-4 py-3 border rounded-lg focus:ring-2 focus:ring-black focus:outline-none ${
+                    errors.password && touched.password ? 'border-red-500' : 'border-gray-300'
+                  }`}
+                  placeholder="Enter your password"
+                />
+              </div>
+              {errors.password && touched.password && (
+                <div className="flex items-center mt-1 text-red-500 text-sm">
+                  <AlertCircle size={16} className="mr-1" />
+                  {errors.password}
+                </div>
+              )}
+            </div>
+
+            <div className="flex items-center justify-between">
+              <label className="flex items-center">
+                <Field
+                  type="checkbox"
+                  name="rememberMe"
+                  className="rounded border-gray-300 text-black focus:ring-black"
+                />
+                <span className="ml-2 text-sm text-gray-600">Remember me</span>
+              </label>
+              <a href="#" className="text-sm text-black hover:underline">
+                Forgot password?
+              </a>
+            </div>
+            
+            <button
+              type="submit"
+              disabled={isSubmitting}
+              className="w-full bg-black text-white py-3 px-4 rounded-lg hover:bg-gray-800 transition-colors flex items-center justify-center space-x-2 text-lg disabled:opacity-50 disabled:cursor-not-allowed"
+            >
+              <span>Sign in</span>
+              <ArrowRight size={20} />
+            </button>
+
+            <div className="text-center text-sm text-gray-600">
+              Don't have an account?{' '}
+              <button
+                type="button"
+                onClick={switchToRegister}
+                className="text-black font-semibold hover:underline"
+              >
+                Sign up
+              </button>
+            </div>
+          </Form>
+        )}
+      </Formik>
+    </div>
+  </div>
+);
+
+export default LoginCard;
